+++ internal/sdkprovider/service/cassandra/cassandra.go
@@ -24,8 +24,10 @@ func cassandraSchema() map[string]*schema.Schema {
 					Computed:    true,
 					Description: "Cassandra server URIs.",
 					Optional:    true,
+					Sensitive:   true,
 					Elem: &schema.Schema{
-						Type: schema.TypeString,
+						Type:      schema.TypeString,
+						Sensitive: true,
 					},
 				},
 			},