+++ internal/sdkprovider/service/clickhouse/clickhouse.go
@@ -23,8 +23,10 @@ func clickhouseSchema() map[string]*schema.Schema {
 					Computed:    true,
 					Description: "ClickHouse server URIs.",
 					Optional:    true,
+					Sensitive:   true,
 					Elem: &schema.Schema{
-						Type: schema.TypeString,
+						Type:      schema.TypeString,
+						Sensitive: true,
 					},
 				},
 			},